Refactor. extracted processing plain text into separate class
