+++ app/src/main/java/com/example/abha_create_verify_android/utils/AadhaarXmlQrParser.kt
@@ -1,25 +1,18 @@
 package com.example.abha_create_verify_android.utils
 
 import android.util.Log
-import org.xml.sax.InputSource
 import org.xmlpull.v1.XmlPullParser
 import org.xmlpull.v1.XmlPullParserException
 import org.xmlpull.v1.XmlPullParserFactory
 import java.io.IOException
 import java.io.StringReader
-import javax.xml.parsers.DocumentBuilderFactory
 
 
-class AadhaarXmlAndTextQr(scanData: String?) {
+class AadhaarXmlQrParser(scanData: String?) {
     private var scannedAadhaarCardInfo: AadhaarCardInfo = AadhaarCardInfo()
 
     init {
-        if(isXmlData(scanData)) {
-            processXmlData(scanData)
-        }
-        else {
-            processPlainTextData(scanData)
-        }
+        processXmlData(scanData)
     }
 
     private fun processXmlData(scanData: String?) {
@@ -105,43 +98,8 @@ class AadhaarXmlAndTextQr(scanData: String?) {
         }
     }
 
-    private fun processPlainTextData(scanData: String?){
-        val keyValuePairs = scanData?.split(", ")
-        val result = mutableMapOf<String, String>()
-
-        if (keyValuePairs != null) {
-            for (pair in keyValuePairs) {
-                if(":" in pair) {
-                    val (key, value) = pair.split(":")
-                    result[key.trim().lowercase()] = value.trim()
-                }
-            }
-        }
-
-        scannedAadhaarCardInfo.name = result[AadhaarDataAttributes.AADHAAR_NAME_ATTR]
-        scannedAadhaarCardInfo.gender = result[AadhaarDataAttributes.AADHAAR_GENDER_ATTR]
-        scannedAadhaarCardInfo.dateOfBirth = result[AadhaarDataAttributes.AADHAAR_DOB_ATTR]
-        scannedAadhaarCardInfo.villageTownCity = result[AadhaarDataAttributes.AADHAAR_VTC_ATTR]
-        scannedAadhaarCardInfo.subDistrict = result[AadhaarDataAttributes.AADHAAR_SUB_DIST_ATTR]
-        scannedAadhaarCardInfo.district = result[AadhaarDataAttributes.AADHAAR_DIST_ATTR]
-        scannedAadhaarCardInfo.state = result[AadhaarDataAttributes.AADHAAR_STATE_ATTR]
-        scannedAadhaarCardInfo.pinCode = result[AadhaarDataAttributes.AADHAAR_PC_ATTR]
-    }
-
     fun getAadhaarCardInfo(): AadhaarCardInfo {
         return scannedAadhaarCardInfo
-    }
-
-    private fun isXmlData(testString: String?): Boolean {
-        return try {
-            val factory = DocumentBuilderFactory.newInstance()
-            val builder = factory.newDocumentBuilder()
-            val inputSource = InputSource(testString?.reader())
-            builder.parse(inputSource)
-            true
-        } catch (e: Exception) {
-            false
-        }
     }
 
 }
